Add xbox xex plugin
This is so you can communicate with the Xbox 360. I don't have the source for this but its public. Feel free to load it in IDA or Ghidra. I'm not the original author.

Load XDRPC.xex as a system boot plugin, load the library in to your project, and enjoy!